sqlarfs: use rowid to read file
Fetch rowid when opening file. Use rowid when reading file.

Note that at this point we do a single read of the whole file, so the
performance improvement should be low. A read by chunks should be
affected.

Note that this naive implementation fetches the rowid when loading a
fs.FileInfo, while it is only used.  The rowid should instead be load
only on Open.